fixup! fixup! fixup! Improve visualization: separate correction plot and fix axis scaling
