+++ src/lerobot/policies/rtc/modeling_rtc.py
@@ -29,7 +29,7 @@ from torch import Tensor
 
 from lerobot.configs.types import RTCAttentionSchedule
 from lerobot.policies.rtc.configuration_rtc import RTCConfig
-from lerobot.policies.rtc.debug_handler import Tracker
+from lerobot.policies.rtc.debug_tracker import Tracker
 
 logger = logging.getLogger(__name__)
